+++ applications/external/weather_station/protocols/emos_e601x.h
@@ -49,7 +49,7 @@ void ws_protocol_decoder_emose601x_feed(void* context, bool level, uint32_t dura
  * @param context Pointer to a WSProtocolDecoderEmosE601x instance
  * @return hash Hash sum
  */
-uint8_t ws_protocol_decoder_emose601x_get_hash_data(void* context);
+uint32_t ws_protocol_decoder_emose601x_get_hash_data(void* context);
 
 /**
  * Serialize data WSProtocolDecoderEmosE601x.
