+++ applications/external/weather_station/protocols/kedsum_th.h
@@ -49,7 +49,7 @@ void ws_protocol_decoder_kedsum_th_feed(void* context, bool level, uint32_t dura
  * @param context Pointer to a WSProtocolDecoderKedsumTH instance
  * @return hash Hash sum
  */
-uint8_t ws_protocol_decoder_kedsum_th_get_hash_data(void* context);
+uint32_t ws_protocol_decoder_kedsum_th_get_hash_data(void* context);
 
 /**
  * Serialize data WSProtocolDecoderKedsumTH.
