+++ applications/external/weather_station/protocols/oregon_v1.h
@@ -49,7 +49,7 @@ void ws_protocol_decoder_oregon_v1_feed(void* context, bool level, uint32_t dura
  * @param context Pointer to a WSProtocolDecoderOregon_V1 instance
  * @return hash Hash sum
  */
-uint8_t ws_protocol_decoder_oregon_v1_get_hash_data(void* context);
+uint32_t ws_protocol_decoder_oregon_v1_get_hash_data(void* context);
 
 /**
  * Serialize data WSProtocolDecoderOregon_V1.
